Blockly: Add labels to array hud
